opendkim/tests/t-sign-rs-tables-bad.conf: change type of signing table
SigningTable is now checked for each entries so that their corresponding
key entry exists in valid form as well as its existence in KeyTable
on loading config file on start up, if the type of SigningTable dataset
supports dkimf_db_walk.

However we need to use bad format entry in KeyTable, to perform the
test, so we use 'refile' dataset type for SigningTable to avoid
the check.